Updated version of MCPEC module from May 2025
The curve analysis was reworked especially for the analysis of double-sigmoid models. The previous version was only able to analyze curves displaying a clear exponential and degradation/death phase. This new version can handle "stacked-double-S-curves" allowing to precisely analyze and quantify different exponential phases indicating metabolic changes.
Due to the enhanced flexibility of the curve analysis, the models can be applied in a much broader context beyond the analysis of biological growth curves.